# fix(dbt-fal): faster inspect stack by using manual solution without loading source code